chore: hacherix feedbacks
