+++ src/repositories/service_providers.py
@@ -15,11 +15,3 @@ class ServiceProvidersRepository:
             return await self.db_session.fetch_one(
                 query, {"service_provider_id": service_provider_id}
             )
-
-    async def get_all(self) -> list[ServiceProviderResponse]:
-        async with self.db_session.transaction():
-            query = """
-            SELECT SP.id, SP.name
-            FROM service_providers as SP
-            """
-            return await self.db_session.fetch_all(query)
