+++ src/routers/auth.py
@@ -1,5 +1,4 @@
 # ------- USER ROUTER FILE -------
-from datetime import timedelta
 
 from fastapi import APIRouter, Depends
 from fastapi.security import OAuth2PasswordRequestForm
@@ -15,8 +14,6 @@ router = APIRouter(
     responses={404: {"description": "Not found"}, 400: {"description": "Bad request"}},
 )
 
-ACCESS_TOKEN_EXPIRE_MINUTES = 60
-
 
 # Add this in your auth router
 @router.post("/token", response_model=Token)
@@ -28,11 +25,8 @@ async def login_for_access_token(
         client_id=form_data.username, client_secret=form_data.password
     )
 
-    # Create access token with the authorized service provider id embedded
-    access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
     access_token = create_access_token(
-        data={"service_provider_id": service_account.service_provider_id},
-        expires_delta=access_token_expires,
+        data={"service_provider_id": service_account.service_provider_id}
     )
 
     return {"access_token": access_token, "token_type": "bearer"}
